Fix “callable classes” compatibility issue for Dart 2
*** BREAKING CHANGE ***
