+++ test/test_components.dart
@@ -1,5 +1,4 @@
 import 'package:react/react.dart';
-import 'package:react/react_client.dart' show ReactComponentFactory, ReactComponentFactoryProxy; // ignore: deprecated_member_use
 
 /// Base component for event handling classes used in test cases.
 class EventComponent extends Component {
@@ -45,9 +44,6 @@ class EventComponent extends Component {
   );
 }
 
-// ignore: deprecated_member_use
-ReactComponentFactory eventComponent = registerComponent(() => new EventComponent()) as ReactComponentFactory;
-
 class SampleComponent extends Component {
   render() => div(props, [
       h1({}, 'A header'),
@@ -57,21 +53,12 @@ class SampleComponent extends Component {
   ]);
 }
 
-// ignore: deprecated_member_use
-ReactComponentFactory sampleComponent = registerComponent(() => new SampleComponent()) as ReactComponentFactory;
-
 class WrapperComponent extends Component {
   render() => div(props, props['children']);
 }
 
-// ignore: deprecated_member_use
-ReactComponentFactory wrapperComponent = registerComponent(() => new WrapperComponent()) as ReactComponentFactory;
+final eventComponent = registerComponent(() => new EventComponent());
 
-final eventComponentV2 = registerComponent(() =>
-    new EventComponent()) as ReactComponentFactoryProxy;
+final sampleComponent = registerComponent(() => new SampleComponent());
 
-final sampleComponentV2 = registerComponent(() =>
-    new SampleComponent()) as ReactComponentFactoryProxy;
-
-final wrapperComponentV2 = registerComponent(() =>
-    new WrapperComponent()) as ReactComponentFactoryProxy;
+final wrapperComponent = registerComponent(() => new WrapperComponent());
